Added haml and sass Converters
The haml converter is based on
    https://gist.github.com/528642/9910ee5d3bbfe94df340be867f55906e077c7508
and Henrik's fork.  Including having Haml wrap attribute values in " instead
of its inane default of '

Added #title and #method_missing to Post so that haml can access them.

Also added #do_setup_once method to Converter, and use it in the new haml
converter along with the markdown, textile and liquid converters to call
the #setup method only once.