Add support for pre_converters and post_converters
These are specified in the yaml data (either as a single name, or array of
names), and those Converters will be run before and after the main converter.

The main converter is still determined dynamically from the file's extension.

This can be useful (for example) to let you still use Liquid tags in an
otherwise haml or markdown page.